+++ app/Http/Controllers/Admin/UserRoleController.php
@@ -2,19 +2,30 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Helpers\Redirect;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
-use Inertia\Inertia;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role as UserRole;
+use Inertia\Inertia;
+use Inertia\Response;
 
 class UserRoleController extends Controller
 {
-    public function index()
+    public function index(): Response | RedirectResponse
     {
-        $roles = UserRole::all();
-        $permissions = Permission::all();
+        if ($this->cannot('viewAny')) {
+            return Redirect::unauthorized();
+        }
+        
+        $roles = UserRole::whereNot('name', 'Super admin')->get();
+
+        $permissions = Permission::orderBy('name')->get()->sortBy(function ($permission) {
+            return strpos($permission, 'order') !== false ? 1 : (strpos($permission, 'user') !== false ? 2 : 3);
+        })->values();
         
         return Inertia::render('Admin/Roles/Index', [
             'roles' => $roles,
@@ -24,9 +35,17 @@ class UserRoleController extends Controller
     
     public function store(Request $request)
     {
-        $request->validate( [
-            'name' => ['required', 'string', 'max:255', 'unique:roles,name'],
-        ]);
+        if ($this->cannot('create')) {
+            return Redirect::unauthorized();
+        }
+
+        $request->validate([
+                'name' => ['required', 'string', 'max:255', 'unique:roles,name', 'regex:/^\S*$/'],
+            ],
+            [
+                'name.regex' => 'The name field cannot contain spaces.',
+            ]
+        );
 
         UserRole::create([
             'name' => $request->name,
@@ -38,9 +57,17 @@ class UserRoleController extends Controller
 
     public function update(Request $request, UserRole $role): RedirectResponse
     {
+        if ($this->cannot('update', $role)) {
+            return Redirect::unauthorized();
+        }
+
         $request->validate([
-            'name' => 'required|string|max:255|unique:roles,name,' . $role->id,
-        ]);
+                'name' => ['required', 'string', 'max:255', 'regex:/^\S*$/', 'unique:roles,name,' . $role->id],
+            ],
+            [
+                'name.regex' => 'The name field cannot contain spaces.',
+            ]
+        );
 
         $role->update([
             'name' => $request->name,
@@ -51,8 +78,21 @@ class UserRoleController extends Controller
 
     public function destroy(UserRole $role): RedirectResponse
     {
+        if ($this->cannot('delete', $role)) {
+            return Redirect::unauthorized();
+        }
+
         $role->delete();
 
         return redirect()->back();
+    }
+
+    private function cannot(string $operation, ?UserRole $role = null): bool
+    {
+        if (!$role) {
+            return Gate::forUser(Auth::guard('admin')->user())->denies($operation, UserRole::class);
+        }
+
+        return Gate::forUser(Auth::guard('admin')->user())->denies($operation, $role);
     }
 }
